+++ src/AccessibilityInsights.Desktop/Misc/StandardLinksHelper.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 
-namespace Axe.Windows.Desktop.Utility
+namespace Axe.Windows.Desktop.Misc
 {
     /// <summary>
     /// Class to access stored link information
