+++ src/AccessibilityInsights.Desktop/UIAutomation/Patterns/SpreadsheetItemPattern.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using Axe.Windows.Core.Bases;
 using UIAutomationClient;
 using Axe.Windows.Core.Attributes;
-using Axe.Windows.Desktop.Utility;
+using Axe.Windows.Desktop.Misc;
 using Axe.Windows.Desktop.Types;
 
 namespace Axe.Windows.Desktop.UIAutomation.Patterns
