+++ src/AccessibilityInsights.Desktop/UIAutomation/Patterns/TextRange.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using UIAutomationClient;
 using Axe.Windows.Core.Attributes;
-using Axe.Windows.Desktop.Utility;
+using Axe.Windows.Desktop.Misc;
 using System.Drawing;
 using System.Runtime.InteropServices;
