+++ src/AccessibilityInsights.DesktopUI/Highlighters/ImageHighlighter.cs
@@ -8,7 +8,7 @@ using System.Windows.Controls;
 using System.Windows.Media;
 using Axe.Windows.Core.Bases;
 using System.Windows.Input;
-using Axe.Windows.Desktop.Utility;
+using Axe.Windows.Desktop.Misc;
 using System;
 using System.Windows.Media.Animation;
 
@@ -100,7 +100,7 @@ namespace AccessibilityInsights.DesktopUI.Highlighters
             // if so make sure that dimension is correct to be visible.
             if(el.PlatformObject == null)
             {
-                var dpi = Axe.Windows.Desktop.Utility.ExtensionMethods.GetDPI((int)Application.Current.MainWindow.Top, (int)Application.Current.MainWindow.Left);
+                var dpi = Axe.Windows.Desktop.Misc.ExtensionMethods.GetDPI((int)Application.Current.MainWindow.Top, (int)Application.Current.MainWindow.Left);
                 this.WinRect = new Rect()
                 {
                     Y = (int)Application.Current.MainWindow.Top,
@@ -229,7 +229,7 @@ namespace AccessibilityInsights.DesktopUI.Highlighters
 
             if (WinRect.IsEmpty)
             {
-                var xyDpi = Axe.Windows.Desktop.Utility.ExtensionMethods.GetWPFWindowPositioningDPI();
+                var xyDpi = Axe.Windows.Desktop.Misc.ExtensionMethods.GetWPFWindowPositioningDPI();
                 this.HighlightWindow.Top = Dimensions.Top / xyDpi;
                 this.HighlightWindow.Left = Dimensions.Left / xyDpi;
