+++ src/AccessibilityInsights.DesktopUI/Highlighters/TextRangeHilighter.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 using Axe.Windows.Core.Enums;
-using Axe.Windows.Desktop.Utility;
+using Axe.Windows.Desktop.Misc;
 using System;
 using System.Collections.Generic;
 using System.Drawing;
