+++ src/AccessibilityInsights.DesktopUI/Highlighters/Win32SnapshotButton.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Desktop.Utility;
-using AccessibilityInsights.DesktopUI.Utility;
+using Axe.Windows.Desktop.Misc;
+using AccessibilityInsights.DesktopUI.Misc;
 using Axe.Windows.Win32;
 using System;
 using System.Collections.Concurrent;
