+++ src/AccessibilityInsights.SharedUx/Utilities/ExtensionMethods.cs
@@ -9,7 +9,7 @@ using Axe.Windows.Actions.Contexts;
 using Axe.Windows.Core.Bases;
 using Axe.Windows.Core.Misc;
 using Axe.Windows.Core.Results;
-using Axe.Windows.Desktop.Utility;
+using Axe.Windows.Desktop.Misc;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
